rag_basics/requirements.txt
python-dotenv==1.1.1
langchain-community==0.3.27
langchain-text-splitters==0.3.8
langchain-google-genai==2.1.6
langchain-qdrant==0.2.0
